Create View
This is the pivotal frontend interface of the Task Manager application, serving as the primary conduit through which users seamlessly engage in a spectrum of CRUD (Create, Read, Update, Delete) operations. With its intuitive design and user-friendly features, this interface empowers users to effortlessly manage tasks, streamlining their workflow and enhancing productivity.